Refactoring http.Client so we can mock requests
By using &http.Client{} directly on our code, it was not possible to
mock external requests from postJson(url) callsite. It was imperative
that we introduced an abstraction so that we could make testing more
painless to work with.

We did that by introducing MockClient struct, which is used as an
interface to interchange the Client type and a global variable GetDoFunc
with the client logic.

With that setup, on future tests and scenarios, we can just set the
desired result for the method http.Client.Do(), so it became more easy
to put our system in a state that we can test and reproduce real
scenarios. What is needed in order to do this is the following:

    GetDoFunc = func(*http.Request) (*http.Response, error) {
        return &http.Response{
            StatusCode: 200,
            Body:       ioutil.NopCloser(bytes.NewReader([]byte(`ok`))),
        }, nil
    }

In the above case, we are returning an http response, but it is possible
also to return an error if that is the scenario you want to test. e.g.:

    GetDoFunc = func(*http.Request) (*http.Response, error) {
        return nil, errors.New("Error from web server")
    }

Co-authored-by: Alexandre Eher <alexandre.eher@werkspot.nl>
Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>