Filter message by tag
Now, only messages with the "just-testing" tag will be processed to be
send as a message to our slack channel. Still, it is hard-coded and
can't be replaced by any configuration. We shoule work next on two
things to make it more flexible. One is the message template. Two is to
find a way to make labels filter configurable. I've some thoughts about
it, as for example, using some url param syntax as in
"Labels=just-testing&User.Name=Eher|Malukenho".

Co-authored-by: Alexandre Eher <alexandre.eher@werkspot.nl>
Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>