Fix docker build
Docker build was failing because we were copying only the "main.go"
file, which is not sufficient anymore, as we decided to work as a
module, and not as a script.

The fix was to sync everything that we have on the current project,
maybe we'll need a dockerignore in the future. But I', not dealing with
it right now unless we have some issue. After wall, code is shipped with
all of it anyway.

Co-authored-by: Alexandre Eher <alexandre.eher@werkspot.nl>
Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>