Treat only the first event
It was necessary to filter out the first event by looking into the
changes that was sent to us through the webhook end-point. We were
receiving all events concerning Merge Request changes, but we actually
needed to notify only when Merge Requests were created.

Now we are checking for the events that have
"changes.createdAt.previous = null", so we know that it is a brand new
Merge Request being open.

Reviewed-by: Emmerson Siqueira <emmersonsiqueira@gmail.com>
Reviewed-by: Alexandre Eher <alexandre.eher@werkspot.nl>
Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>