Just notify when new merge requests are created
We were creating notifications for every action related to
merge_request.
Is can be too verbose for a Slack Channel.

For now we are looking just for:

	EventType = "merge_request"
	ObjectAttributes.Action = "open"

We have plans to update the previous Slack message instead of notify
again.

With that we are incresing the code coverage from 78.7% to 87.2%

Co-authored-by: Jefersson Nathan <malukenho.dev@gmail.com>
Signed-off-by: Alexandre Eher <alexandre.eher@werkspot.nl>